Untangle reverse dependencies
For the longest time we had a slight mess of reverse dependencies
between telamon itself and the backend crates (cuda and mppa,
specifically).   This was due to using the cuda crate in some tests, and
for historical reasons that are no longer relevant for the mppa crate
(the corresponding code is now dead anyways).

This patch untangles the mess by:

 - Moving `tests/cuda.rs` into the cuda backend crate (and running tests
   for the cuda backend crate in travis)

 - Moving the `matmul` example to `telamon-cli` (this one might be
   controversial)

 - Removing the `total_num_threads` function from the `NameMap`

 - Finally the `mppa` and `cuda` features from the `telamon` crate are
   removed
