+++ .travis.yml
@@ -29,6 +29,7 @@ jobs:
     - stage: test
       script: CUDA_HOME=/usr/local/cuda-10.1 cargo check --all --exclude telajax --all-targets --tests --benches
     - script: cargo test
+    - script: cd backend/cuda && cargo test
     - script: cd telamon-utils && cargo test
     - script: cd telamon-gen && cargo test
     - script: cd telamon-gen/cc_tests && cargo test
