+++ telamon-cli/examples/matmul.rs
